Core/Units: moved health and power ordering predicates from Unit header into CommonPredicates
Port From (https://github.com/TrinityCore/TrinityCore/commit/2f6ed2c203b16a1d1e85f61a8b8e2cf3d1a4e784)

+++ Source/Game/Entities/Pet.cs
@@ -1497,7 +1497,7 @@ namespace Game.Entities
 
         public override bool IsLoading() { return m_loading; }
 
-        public override byte GetPetAutoSpellSize() { return (byte)m_autospells.Count; }
+        public override int GetPetAutoSpellSize() { return m_autospells.Count; }
         public override uint GetPetAutoSpellOnPos(byte pos)
         {
             if (pos >= m_autospells.Count)
